WIP on SecurityPluginAuthenticationManager
