+++ sandbox/libs/authn/licenses/log4j-NOTICE.txt
@@ -1,0 +1,5 @@
+Apache log4j
+Copyright 2007 The Apache Software Foundation
+
+This product includes software developed at
+The Apache Software Foundation (http://www.apache.org/).
