+++ sandbox/libs/authn/licenses/log4j-api-2.17.1.jar.sha1
@@ -1,0 +1,1 @@
+d771af8e336e372fb5399c99edabe0919aeaf5b2
