--- sandbox/libs/authn/src/main/java/org/opensearch/authn/securityplugin/SecurityPluginAccessTokenManager.java
@@ -1,29 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-package org.opensearch.authn.securityplugin;
-
-import org.opensearch.authn.AccessTokenManager;
-import org.opensearch.authn.tokens.AccessToken;
-
-public class SecurityPluginAccessTokenManager implements AccessTokenManager {
-    @Override
-    public void expireAllTokens() {
-        return;
-    }
-
-    @Override
-    public AccessToken generate() {
-        return null;
-    }
-
-    @Override
-    public AccessToken refresh(AccessToken token) {
-        return null;
-    }
-}
