+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/SecurityRestFilter.java
@@ -12,6 +12,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.shiro.authc.AuthenticationException;
 import org.opensearch.OpenSearchException;
+import org.opensearch.authn.Identity;
 import org.opensearch.authn.Subject;
 import org.opensearch.identity.jwt.JwtVendor;
 import org.opensearch.authn.tokens.AuthenticationToken;
