+++ sandbox/modules/identity/src/test/java/org/opensearch/identity/IdentityTests.java
@@ -6,6 +6,7 @@
 package org.opensearch.identity;
 
 import org.opensearch.authn.AuthenticationManager;
+import org.opensearch.authn.Identity;
 import org.opensearch.test.OpenSearchTestCase;
 
 import static org.mockito.Mockito.mock;
